+++ lib/bin/watch.ts
@@ -1,5 +1,5 @@
 import { ChildProcess } from 'child_process'
-import CONFIG from '../config/config'
+import CONFIG from '../runtime/config/config'
 import clean from './clean'
 import build from './build'
 import start from './start'