+++ lib/config/webpack.client.prod.ts
@@ -1,7 +1,7 @@
 import * as webpack from 'webpack'
 import * as merge from 'webpack-merge'
 import * as ExtractTextPlugin from 'extract-text-webpack-plugin'
-import CONFIG from './config'
+import CONFIG from '../runtime/config/config'
 import baseConfig from './webpack.base'
 import clientBaseConfig from './webpack.client'
 import prodConfig from './webpack.prod'