+++ lib/index.ts
@@ -12,6 +12,7 @@ import binWatch from './bin/watch'
 import * as buildModule from './build'
 import * as configModule from './config'
 import * as lintModule from './lint'
+import * as runtimeModule from './runtime'
 import * as serverModule from './server'
 import * as statsModule from './stats'
 
@@ -46,6 +47,7 @@ export const bin: BinType = {
 export const build = buildModule
 export const config = configModule
 export const lint = lintModule
+export const runtime = runtimeModule
 export const server = serverModule
 export const stats = statsModule