+++ lib/runtime/server/assets.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs'
 import * as path from 'path'
 import CONFIG from '../config/config'
-import { Assets } from '../types'
+import { Assets } from '../../types'
 
 const root = process.cwd()
 const assetsFile = path.resolve(root, 'assets.json')