+++ lib/runtime/server/server.ts
@@ -1,7 +1,9 @@
+import * as fs from 'fs'
+import * as path from 'path'
 import * as express from 'express'
-import { getDefaultHtmlMiddleware, getHotReloadMiddleware, openBrowser } from './dev'
+import { addAssetsToHtml } from './assets'
 import { findFreePort } from '../util/network'
-import { log } from '../util/log'
+import { log, logError } from '../util/log'
 import CONFIG from '../config/config'
 
 const isProduction = process.env.NODE_ENV === 'production'
@@ -19,6 +21,34 @@ export const isFastMode = () => (
     process.env.START_FAST_MODE === 'true'
 )
 
+export const expressNoop: express.RequestHandler = (_res, _req, next) => next()
+
+export const getDefaultHtmlMiddleware = (logNotFound = false) => {
+    let indexContent: string
+
+    if (SERVER_PUBLIC_DIR) {
+        try {
+            const indexPath = path.resolve(SERVER_PUBLIC_DIR, 'index.html')
+            indexContent = fs.readFileSync(indexPath, 'utf8')
+        } catch (e) {
+            if (logNotFound) {
+                logError('Reading index.html failed!', e)
+            }
+            return expressNoop
+        }
+    }
+
+    const middleware: express.RequestHandler = (_req, res) => {
+        const indexWithAssets = addAssetsToHtml(indexContent)
+        res
+            .status(200)
+            .contentType('text/html')
+            .send(indexWithAssets)
+    }
+
+    return middleware
+}
+
 export type CreateServerType = (
     middlewareHook?: (app: express.Express) => void,
     callback?: () => void,
@@ -32,6 +62,8 @@ export const createServer: CreateServerType = (
     const app = express()
 
     if (!isProduction && isWatchMode()) {
+        // tslint:disable-next-line no-var-requires
+        const { getHotReloadMiddleware } = require('../../server/dev')
         app.use(getHotReloadMiddleware())
     }
 
@@ -51,6 +83,8 @@ export const createServer: CreateServerType = (
         const server = app.listen(usePort, () => {
             log(`Server listening on port ${usePort}`)
             if (!isProduction && isWatchMode()) {
+                // tslint:disable-next-line no-var-requires
+                const { openBrowser } = require('../../server/dev')
                 openBrowser(usePort)
             }
             if (callback) {